pyproject.toml
[tool.poetry]
name = "kompose"
version = "0.5.0"
description = ""
authors = ["Jefte Keller <kellergnr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "==3.3.4"
Django = "==3.2"
djangorestframework = "==3.12.4"
pytz = "==2021.1"
sqlparse = "==0.4.1"
coverage = "^5.5"

[tool.poetry.dev-dependencies]
black = {version = "^21.6b0", allow-prereleases = true}
pylint = "^2.9.3"
flake8 = "^3.9.2"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
